@@ -47,6 +47,7 @@
     "glob": "5.0.3",
     "htmlparser2": "^3.8.3",
     "imports-loader": "0.6.5",
+    "invariant": "2.2.1",
     "less": "2.3.1",
     "less-loader": "2.0.0",
     "lodash": "3.10.1",
@@ -54,7 +55,8 @@
     "multiline-slash": "2.0.0",
     "normalize.css": "^3.0.3",
     "raw-loader": "0.5.1",
-    "react": "^0.13.3",
+    "react": "15.0.2",
+    "react-dom": "15.0.2",
     "react-transclusion": "1.0.0-1",
     "style-loader": "0.13.1",
     "urijs": "1.18.0",
@@ -69,7 +71,6 @@
     "eslint": "^1.4.3",
     "eslint-plugin-react": "^3.4.1",
     "highlight.js": "^8.8.0",
-    "history": "1.17.0",
     "html": "0.0.10",
     "istanbul": "^1.0.0-alpha.2",
     "jquery": "2.2.3",
@@ -82,9 +83,8 @@
     "karma-webpack": "1.7.0",
     "mocha": "^2.3.3",
     "mocha-lcov-reporter": "^1.0.0",
-    "react-drill": "1.0.5",
+    "react-drill": "2.1.1",
     "react-hot-loader": "1.3.0",
-    "react-router": "0.13.x",
     "serve-static": "^1.10.0",
     "sinon": "^1.17.1",
     "webpack-dev-middleware": "1.6.1",
