@@ -2,7 +2,7 @@ var write = require('./HTMLSerializer__write');
 var emitFiles = require('./HTMLSerializer__emitFiles');
 var async = require('async');
 var b = require('tinydoc-corpus').Types.builders;
-var invariant = require('react/lib/invariant');
+var invariant = require('invariant');
 
 module.exports = function(config) {
   var exports = {};
