@@ -102,18 +102,24 @@ function HashBasedURIGenerator(/*config*/) {
       return ensureLeadingHash(node.meta.href);
     }
 
-    if (node.type === 'DocumentEntity') {
-      return NodeURI(node.parentNode) + '/' + encodeURI(node.id);
-    }
-    else if (node.type === 'Document') {
-      return NodeURI(node.parentNode) + '/' + encodeURI(node.id);
-    }
-    else if (node.type === 'Namespace') {
-      return '#/' + encodeURI(node.id);
-    }
-    else if (node.type === 'Corpus') {
+    if (node.type === 'Corpus') {
       return null;
     }
+
+    return '#/' + encodeURI(node.uid);
+
+    // if (node.type === 'DocumentEntity') {
+    //   return NodeURI(node.parentNode) + '/' + encodeURI(node.id);
+    // }
+    // else if (node.type === 'Document') {
+    //   return NodeURI(node.parentNode) + '/' + encodeURI(node.id);
+    // }
+    // else if (node.type === 'Namespace') {
+    //   return '#/' + encodeURI(node.id);
+    // }
+    // else if (node.type === 'Corpus') {
+    //   return null;
+    // }
   }
 
   function NodeAnchor(node) {
