@@ -3,7 +3,7 @@ var assign = require('lodash').assign;
 var fs = require('fs');
 var path = require('path');
 var K = require('./HTMLSerializer__constants');
-var invariant = require('react/lib/invariant');
+var invariant = require('invariant');
 
 function buildRelativeAssetList(list, distanceFromRoot) {
   if (distanceFromRoot < 1) {
