@@ -1,9 +1,29 @@
 const Subject = require("../Article");
+const React = require('react');
 const reactSuite = require("test_helpers/reactSuite");
 const { assert } = require('chai');
+const { object, } = React.PropTypes;
 
 describe("Markdown::Components::Article", function() {
-  reactSuite(this, Subject, {
+  reactSuite(this, React.createClass({
+    childContextTypes: {
+      location: object,
+      config: object,
+    },
+
+    getChildContext() {
+      return {
+        location: {},
+        config: {
+          disqus: false
+        }
+      }
+    },
+
+    render() {
+      return <Subject {...this.props} />
+    }
+  }), {
     documentNode: {
       meta: {},
       properties: {
