@@ -1,141 +1,144 @@
-const Subject = require('../Link');
+const Link = require('../Link');
 const React = require('react');
+const ReactDOM = require('react-dom');
 const sinonSuite = require('test_helpers/sinonSuite');
 const { assert } = require('chai');
 const CorpusAPI = require('core/CorpusAPI');
 const AppState = require('core/AppState');
 const { drill } = require('react-drill');
+const { string, func, shape, object, } = React.PropTypes;
+const { assign } = require('lodash');
 
-let originalLocation;
 
 describe('tinydoc::Components::Link', function() {
+  let linkContext;
   const sinon = sinonSuite(this);
+  const Subject = React.createClass({
+    childContextTypes: {
+      config: object,
+      navigate: func,
+      location: shape({
+        pathname: string,
+      })
+    },
+
+    getChildContext() {
+      return linkContext;
+    },
+
+    render() {
+      return <Link {...this.props} />
+    }
+  });
 
   let subject;
   let container;
 
   beforeEach(function() {
-    originalLocation = window.location;
+    linkContext = null;
     container = document.createElement('div');
   });
 
   afterEach(function() {
-    restoreLocation();
-
-    React.unmountComponentAtNode(container);
+    linkContext = null;
+    ReactDOM.unmountComponentAtNode(container);
     container.remove();
     container = null;
   });
 
   context('given @href', function() {
     it('uses it without any questions asked', function() {
-      subject = React.render(<Subject href="/foobar.js" />, container);
+      render({ href: "/foobar.js" });
+
       assert.include(drill(subject).node.href, '/foobar.js');
     });
 
-    context('in SinglePageMode...', function() {
-      beforeEach(function() {
-        sinon.stub(AppState, 'inSinglePageMode', () => true);
+    it('builds a relative URL', function() {
+      render({ href: '/foo/bar.html' }, {
+        location: {
+          pathname: '/foo/index.html'
+        }
       });
 
-      it('should use @anchor as an anchor, if present', function() {
-        render({ href: '/foo.html', anchor: 'foo' });
-
-        assert.include(drill(subject).node.href, '/foo.html#foo');
-      });
-
-      it('should use the anchor found in @href, otherwise', function() {
-        render({ href: '/foo.html#foo' });
-
-        assert.include(drill(subject).node.href, '/foo.html#foo');
-      });
-
-      // it('should complain if it finds no anchor', function() {
-      //   sinon.stub(console, 'warn');
-
-      //   render({ href: '/foo.html' });
-
-      //   assert.calledWith(console.warn, sinon.match(/An anchor is required/));
-      //   assert.include(drill(subject).node.href, '/foo.html');
-      // });
+      assert.equal(drill(subject).node.href.replace(location.origin + '/', ''), 'bar.html');
     });
   });
 
-  context('given @to that points to a corpus node...', function() {
+  context('in SinglePageMode...', function() {
+    const config = { layoutOptions: { singlePageMode: true } };
+
+    context('given @to...', function() {
+      let corpus;
+
+      beforeEach(function() {
+        corpus = buildSPMCorpus();
+      });
+
+      it('should use @meta.href', function() {
+        render({ to: corpus.get('lua') });
+
+        assert.include(drill(subject).node.href, '#/lua');
+      });
+
+      it('should link to entities using @meta.href', function() {
+        render({ to: corpus.get('lua/from_lua') });
+
+        assert.include(drill(subject).node.href, '#/lua/from_lua');
+      });
+
+      it('should link to entities with hashes inside their URIs', function() {
+        render({ to: corpus.get('lua/cli#set_description') });
+
+        assert.include(drill(subject).node.href, '#/lua/cli#set_description');
+      });
+    });
+  });
+
+  context('given @to...', function() {
     let corpus;
 
     beforeEach(function() {
       corpus = buildCorpus();
-      subject = React.render(<Subject to={corpus.get('api/foo')} />, container);
     });
 
-    it('links to the node', function() {
-      assert.include(drill(subject).node.href, '/api/foo.html');
-    });
-
-    context('in SinglePageMode...', function() {
-      beforeEach(function() {
-        sinon.stub(AppState, 'inSinglePageMode', () => true);
-      });
-
-      it('should use @meta.anchor', function() {
-        render({ to: corpus.get('api/foo') });
-
-        assert.include(drill(subject).node.href, '/foo.html#foo');
-      });
+    it('links to the node using its @href', function() {
+      render({ to: corpus.get('api/Bag') });
+      assert.include(drill(subject).node.href, '/api/Bag.html');
     });
   });
 
   context('when using file:// as a protocol', function() {
-    beforeEach(function() {
-      stubLocation({
-        protocol: 'file:'
-      });
-    });
-
     it('renders', function() {
-      subject = React.render(<Subject href="/api.html" />, container);
+      render({ href: "/api.html" }, {
+        location: {
+          protocol: 'file:',
+          pathname: '/srv/http/docs/tinydoc-test/index.html'
+        }
+      });
 
       assert.ok(subject.isMounted());
     });
   });
 
+  function render(props, context) {
+    linkContext = buildContextWithDefaults(context);
+    subject = ReactDOM.render(<Subject {...props} />, container);
+  }
 
-  function render(props) {
-    subject = React.render(<Subject {...props} />, container);
+  function buildContextWithDefaults(context = {}) {
+    return assign({
+      navigate: sinon.stub(),
+      location: assign({
+        pathname: '/'
+      }, context.location),
+    }, context);
   }
 });
 
-
-function stubLocation(newLocation) {
-  // Object.defineProperty(window, 'location', {
-  //   configurable: true,
-  //   enumerable: true,
-  //   writable: true,
-  //   value: newLocation
-  // });
-}
-
-function restoreLocation() {
-  // window.location = originalLocation;
-}
-
 function buildCorpus() {
-  return CorpusAPI({
-    'api': {
-      type: 'Namespace',
-      meta: {
-        href: null
-      },
-      documents: [ 'api/foo' ],
-    },
-    'api/foo': {
-      type: 'Document',
-      parentNode: 'api',
-      meta: {
-        href: '/api/foo.html',
-        anchor: 'foo'
-      }
-    }
-  });
+  return CorpusAPI(require('json!test_helpers/fixtures/corpus'));
+}
+
+function buildSPMCorpus() {
+  return CorpusAPI(require('json!test_helpers/fixtures/corpus--spm'));
 }
